// frontend/src/types/types.ts
export interface Attendee {
    student_id: string; 
    name: string;
    email?: string; 
    sex: string;
    religion?: string; 
    nation?: string; 
    phone_number?: string; 
    class?: string; 
    position: string;
    type_degate: string;
    checkin: boolean;
    checkout: boolean;
  }
  